fix problems related to 64-bit compilation in lib
. dbghelp is now pulled in by a separate external_libraries header
. disable _SECURE_SCL (improves perf)
. amd64: initial implementation of essential functions, not yet complete
. manifest: disabled on ICC because its IPO module incorrectly interprets /manifestdependency
. wdbg_sym: no longer support stack walks on non-ia32 win2k (avoids trouble due to mixing SEH and dtors)
. winit and wstartup: had to remove forced symbol include - does not link otherwise on ICC (for reasons unknown). function pointers are actually included and called on x64 windows (compiling with MSC)

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5895 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/allocators/headerless.cpp
@@ -360,7 +360,7 @@ private:
 
 	// bit i set <==> size class i's freelist is not empty.
 	// this allows finding a non-empty list in O(1).
-	uintptr_t m_bitmap;
+	u32 m_bitmap;
 };
